requirements.txt
livekit-agents>=1.0.0
assemblyai>=0.21.0
python-dotenv>=1.0.0
groq>=0.4.0
openai>=1.0.0
asyncio
logging
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
websockets>=11.0.0
aiohttp>=3.8.0
python-multipart>=0.0.6
railway>=0.1.0 